+++ litebridge-maven-plugin/src/main/java/org/litebridge/maven/config/reverse/RevEngOutputConfig.java
@@ -1,5 +1,6 @@
 package org.litebridge.maven.config.reverse;
 
+import org.apache.maven.plugins.annotations.Parameter;
 import org.jspecify.annotations.Nullable;
 import org.litebridge.maven.config.OutputConfig;
 
@@ -13,6 +14,19 @@ public class RevEngOutputConfig extends OutputConfig {
      */
     private @Nullable RevEngJSpecifyConfig jspecify;
 
+    /**
+     * Controls how related entities are mapped when foreign keys are encountered.
+     * <p>
+     * Setting it to {@code false} disables entity resolution; foreign key fields will use the "flat" database type
+     * (e.g. {@code Long personId}) instead of replacing the field with the related entity (e.g. {@code Person person}.
+     * <p>
+     * This is the global setting for all tables. It can be overridden on a per-table basis.
+     * <p>
+     * Default: {@code true}
+     */
+    @Parameter(defaultValue = "true")
+    private boolean resolveRelationships = true;
+
     @Override
     public void setJspecify(final @Nullable RevEngJSpecifyConfig jspecify) {
         super.setJspecify(jspecify);
@@ -21,5 +35,13 @@ public class RevEngOutputConfig extends OutputConfig {
     @Override
     public @Nullable RevEngJSpecifyConfig getJspecify() {
         return super.getJspecify();
+    }
+
+    public boolean isResolveRelationships() {
+        return resolveRelationships;
+    }
+
+    public void setResolveRelationships(final boolean resolveRelationships) {
+        this.resolveRelationships = resolveRelationships;
     }
 }