+++ CHANGELOG.md
@@ -3,10 +3,10 @@
 ## 0.3.2
 
 ### Added
-- None.
+- Maven plugin: Added "resolveRelationships" config parameter to control foreign key behaviour.
 
 ### Changed
-- Fixed issue with complex circular dependencies in the Litebridge Maven plugin's `reverse-engineer` goal.
+- Maven plugin: Fixed issue with complex circular dependencies in the Litebridge Maven plugin's `reverse-engineer` goal.
 
 ### Deprecated
 - None.